requirements.txt
Librarys:
    Pillow
    String
    Random

Fonts:
    Melinda script.ttf
    RobotoMono-Thin.ttf